+++ odbc-api/src/cursor.rs
@@ -1,10 +1,9 @@
 use crate::{
+    buffers::Indicator,
     handles::{Statement, StatementImpl},
-    parameter::VarCharMut,
+    parameter::VarCharSliceMut,
     ColumnDescription, DataType, Error, Output,
 };
-
-use odbc_sys::{NO_TOTAL, NULL_DATA};
 
 use std::{
     borrow::BorrowMut,
@@ -167,45 +166,46 @@ where
         // We repeatedly fetch data and add it to the buffer. The buffer length is therefore the
         // accumulated value size. This variable keeps track of the number of bytes we added with
         // the current call to get_data.
-        let mut fetch_size: isize = buf.len().try_into().unwrap();
-        let mut target = VarCharMut::from_buffer(buf.as_mut_slice(), None);
+        let mut fetch_size = buf.len();
+        let mut target = VarCharSliceMut::from_buffer(buf.as_mut_slice(), Indicator::Null);
         // Fetch binary data into buffer.
         self.get_data(col_or_param_num, &mut target)?;
         let not_null = loop {
             match target.indicator() {
                 // Value is `NULL`. We are done here.
-                NULL_DATA => {
+                Indicator::Null => {
                     buf.clear();
                     break false;
                 }
                 // We do not know how large the value is. Let's fetch the data with repeated calls
                 // to get_data.
-                NO_TOTAL => {
+                Indicator::NoTotal => {
                     let old_len = buf.len();
                     // Use an exponential strategy for increasing buffer size. +1 For handling
                     // initial buffer size of 1.
                     buf.resize(old_len * 2, 0);
-                    target = VarCharMut::from_buffer(&mut buf[(old_len - 1)..], None);
+                    target =
+                        VarCharSliceMut::from_buffer(&mut buf[(old_len - 1)..], Indicator::Null);
                     self.get_data(col_or_param_num, &mut target)?;
                 }
                 // We did get the complete value, including the terminating zero. Let's resize the
                 // buffer to match the retrieved value exactly (excluding terminating zero).
-                indicator if indicator < fetch_size => {
-                    assert!(indicator >= 0);
+                Indicator::Length(len) if len < fetch_size => {
                     // Since the indicator refers to value length without terminating zero, this
                     // also implicitly drops the terminating zero at the end of the buffer.
-                    let shrink_by: usize = (fetch_size - indicator).try_into().unwrap();
+                    let shrink_by = fetch_size - len;
                     buf.resize(buf.len() - shrink_by, 0);
                     break true;
                 }
                 // We did not get all of the value in one go, but the data source has been friendly
                 // enough to tell us how much is missing.
-                indicator => {
-                    let still_missing = (indicator - fetch_size + 1) as usize;
-                    fetch_size = (still_missing + 1).try_into().unwrap();
+                Indicator::Length(len) => {
+                    let still_missing = len - fetch_size + 1;
+                    fetch_size = still_missing + 1;
                     let old_len = buf.len();
                     buf.resize(old_len + still_missing, 0);
-                    target = VarCharMut::from_buffer(&mut buf[(old_len - 1)..], None);
+                    target =
+                        VarCharSliceMut::from_buffer(&mut buf[(old_len - 1)..], Indicator::Null);
                     self.get_data(col_or_param_num, &mut target)?;
                 }
             }